+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 0.1.4
-- Fix: `DebugConfigurationProvider` should set a default only for required values for the specific configuration type. 
+- Fix: `DebugConfigurationProvider` should set a default only for required values for the specific configuration type.
+- Filter out empty or unsupported-type skip-paths.
 
 ## 0.1.3
 - Implement `skip_path` support based on a custom DAP request.